Fixed #104 - cluster/hostnames with hyphen.
But also uncovered that the technique of linking cluster to host by hostbame and/or hostname user alias is not fool proof and has issues if anything other than lowercase is used. This does not seem to be resolvable currently without chages to the Atlas API itself, since there is no direct link between cluster and host in the endpoints.